refactor(game/ui): move GameCardSkeleton to entity layer
