+++ service/app/src/entities/game/ui/GameCard/GameCardSkeleton.tsx
@@ -1,0 +1,8 @@
+export const GameCardSkeleton = () => {
+  return (
+    <div className="flex w-[178px] flex-col items-start gap-[14px]">
+      <div className="size-[178px] animate-pulse rounded-[10px] bg-gray-200" />
+      <div className="h-[46px] w-[178px] animate-pulse rounded bg-gray-200" />
+    </div>
+  )
+}
