+++ service/app/src/entities/game/ui/GameCard/gameCard.tsx
